apiconnect/scripts/requirements.txt
pyyaml>=6.0
requests>=2.26.0